Move ftdetect to ftplugin and rename script
Renames the ftdetect folder to ftplugin and renames neuron.vim to
markdown.vim. This is more correct as the neuron.vim script *did not*
actually set the filetype using `setfiletype` and instead just set up
some autocmds. This allows for markdown plugins to be installed and
functioning as expected while also fully supporting this neuron plugin