Don't replace words when editting a new zettel
This was causing some strange behavior when creating a new zettel from a
selection, as it would change the character under the cursor after
deleting the selection. This was usually a space, but in words that were
followed by punctuation, it would end up replacing the punctuation
character. Instead, we now just insert the text using `i`.